Database resilience (#143)
Make secure message start up resilient to slow db start up

Retry database connection on DatabaseError every 2s for up to 30s. This
allows the app to be more robust when starting up. For instance when
running up apps via docker-compose the app should be able to recover if
the database is not up first.