Improve logging
Previously an exception would be raised, but nothing would be logged and you would have to
look at logs of other applications to get a hint of what went wrong.  Now we catch the exception
log an error with the same error text, then raise the exception again.
This way the functionality doesn't change, and we get better logging!