Make all changes then commit.
Previously it would save the event, then amend the message, with commits between these two actions.
If anything went wrong then the database could be left in an incorrect state.  This change fixes that by
making all the changes first, then commiting them all at the same time (so they can all be rolled back in
the event of an error)