Fix get_user_details in party service
get_user_details used to take a single value.  It was then changed to take a list...except some things were
still giving it a single value, and when you do a for loop over a string, it operates on each character of the string
instead of the whole thing, leading to some very incorrect calls being made to the party service (i.e., instead of
id=a-1c, it would do id=a&id=-&id=1&id=c).